Removed code line removing duplicates from warning
In a previous version duplicate messages from Nmap stderr got removed. As we understand these messages as warnings now, I think it's not good to alter/clean these messages anymore, in order to give full control to the developer. The developer still could do a "unique" on the returned list of messages on his own. In other cases he might want to know how often a certain warning happened, etc...